mingw-w64 build fix.
This is kludgy, but it's a quick fix and it works for both old versions and
updated versions of mingw-w64. It won't fix the build problems for anyone
with an updated version of xinput.h and older versions of GCC, but pacman -Syu
should take care of that on its own.