Fix undefined symbols in Linux builds.
Add core/unix/ and joystick/steam/ to the sources in Linux (since they are
referred to from core/linux/ and joystick/linux/, respectively). Also add
the missing HAVE_TAN and HAVE_TANF to the prebuilt SDL_config_linux.h, so
we don't use but fail to compile in the uclibc version.