Try another ugly workaround for the xinput thing
I don't see a really general way to handle this without a compile check.
Since we aren't using a real build system, we don't have much
infrastructure for compile checks, and so I just implemented it with a
shell script. This is pretty ugly, though less ugly than the last
attempt.

For an example of how to properly do this compile check, see
e.g. CMakeLists.txt.