Fix mac resizing behavior
When a window is programmatically resized on a mac, the OS doesn't
necessarily give the requested dimensions, and in particular, if the
height would not fit between the menubar and a dock on the bottom of the
screen, it will only give the max usable height. SDL was keeping in its
state only the requested window size, not the window size actually
chosen by the OS, and this was causing a bunch of problems. (For one
thing, since `SDL_GL_GetDrawableSize` has the *correct* value, our
calculations of device density were off.) One manifestation of this bug
is: https://crawl.develz.org/mantis/view.php?id=11895

I have also submitted an upstream bug report at:
https://bugzilla.libsdl.org/show_bug.cgi?id=4574