Fix the build on M1 ARM (12407)
This is to make sure that `OSAtomicCompareAndSwap32Barrier` is
explicitly defined. I'm not actually sure that this is the right
solution in general -- it might be that this should instead use the 64
bit version of this call on current macs. However, this bit of SDL isn't
used by crawl, so it doesn't matter very much.