fix: merge upstream changes to capslock handling
This is aimed to prevent MacOS builds (post catalina) from asking for
input monitoring permission; it's unnecessarily alarming and isn't
needed. Plus, as far as I can tell the permissions request resulted from
a misimplementation of capslock handling in the first place.

This change was originall made in
https://github.com/libsdl-org/SDL/commit/57e08c27ef4a54e48c20c6bfd7c841216a6b70b5,
but I've taken up-to-date SDL2 code, which seems to have some other
tweaks.